requirements.txt
streamlit>=1.24.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.14.0
folium>=0.15.0
streamlit-folium>=0.15.0
gdown>=4.7.1
python-dotenv>=1.0.0

# Dependencias para análisis de datos
matplotlib>=3.8.0
seaborn>=0.13.0
openpyxl>=3.1.5
ipykernel>=6.0.0  # Para Jupyter notebooks

# Dependencias para servicios de Google Cloud
google-cloud-bigquery>=3.11.0
google-cloud-bigquery-storage>=2.22.0
google-cloud-storage>=2.10.0  # Para el sistema de feedback
pyarrow>=14.0.1  # Requerido para optimizar el rendimiento de BigQuery
